feat: JWT에서 멤버 ID 추출하는 기능 추가
