--- backend/src/main/java/com/now/naaga/auth/damain/jwt/JwtProvider.java
@@ -1,30 +1,0 @@
-package com.now.naaga.auth.damain.jwt;
-
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.io.Decoders;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import java.security.Key;
-import java.util.Date;
-
-@Component
-public class JwtProvider {
-
-    private final Key key;
-
-    public JwtProvider(@Value("${jwt.secret-key}") final String secretKey) {
-        byte[] keyBytes = Decoders.BASE64.decode(secretKey);
-        this.key = Keys.hmacShaKeyFor(keyBytes);
-    }
-
-    public String generate(String subject, Date expiredAt) {
-        return Jwts.builder()
-                .setSubject(subject)
-                .setExpiration(expiredAt)
-                .signWith(key, SignatureAlgorithm.HS512)
-                .compact();
-    }
-}
