+++ backend/src/main/java/com/now/naaga/auth/infrastructure/jwt/JwtGenerator.java
@@ -1,4 +1,4 @@
-package com.now.naaga.auth.damain.jwt;
+package com.now.naaga.auth.infrastructure.jwt;
 
 import com.now.naaga.auth.damain.AuthTokens;
 import org.springframework.stereotype.Component;
